pyproject.toml
[tool.poetry]
name = "karin"
version = "0.1.0"
description = "IwataGumi Discord Bot"
authors = ["IwataGumi Member"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.8.10"
discord = {extras = ["voice"], version = "^2.3.1"}
pynacl = "^1.5.0"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"
mecab-python3 = "^1.0.6"
unidic-lite = "^1.0.8"
alembic = "^1.11.1"
sqlalchemy = "^2.0.18"
psycopg2 = "^2.9.6"
sqlalchemy-utils = "^0.41.1"
requests = "^2.31.0"
pyyaml = "^6.0"
pyi18n-v2 = "^1.2.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.3.0"
ipykernel = "^6.24.0"
notebook = "^6.5.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
